move FollowedHashtagItem to :core:commonui:content
